mcp-servers/knowledge-mcp/requirements.txt
# FastMCP Framework (replaces mcp==1.0.0)
fastmcp

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4

# HTTP requests for Ollama
requests==2.31.0

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0 